refactor common classes
