+++ TeamCode/build.gradle
@@ -18,6 +18,10 @@ apply from: '../build.dependencies.gradle'
 android {
     namespace = 'org.firstinspires.ftc.teamcode'
 
+    defaultConfig {
+        versionName "2.0"
+    }
+
     packagingOptions {
         jniLibs.useLegacyPackaging true
     }
