+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleOp/ControlsNEW.java
@@ -1,7 +1,9 @@
-package org.firstinspires.ftc.teamcode.TeleOp;
+package org.firstinspires.ftc.teamcode.teleOp;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+
+import org.firstinspires.ftc.teamcode.commons.RobotHardware;
 
 @TeleOp
 public class ControlsNEW extends LinearOpMode {
